Change alarm descriptions to link to ops manual
When an alarm (aka alert) fires, it is useful context for the people
responding to each alarm to have contextual documentation about how to
respond.

This change alters the `AlarmDescription` field of each
`AWS::CloudWatch::Alarm` to link to the relevant part of our operations
manual.

The operations manual has been broken out of a single Word document into
many small focused documents to allow this direct linking. Directly
linking to a section with a Sharepoint document is difficult (if
possible?) and fragile.

I’ve also renamed RDSQueryLatency to be RDSModifyLatency. It wasn’t
query (or read) latency. It was the latency associated with mutation
operations (insert, update, and delete).

I didn’t name it WriteLatency since there is already an existing RDS WriteLatency metric which is related to IO, rather than logical SQL
operations.

I also added an alarm to check SelectLatency on the database. For
similar reasons, I didn’t name this ReadLatency due to the existing RDS
IO ReadLatency metric.

Final change is to use files to pass in parameters, rather than the CLI.

It makes it easier for future operators, and does not appear to have
any security implications at this time.